Data Analysis Robusta price
Robusta Coffee

The main drivers influencing the price of coffee are demand and supply. In other words, demand for coffee beans is determined by companies such as Starbucks, Nestle, and Kenko. Consequently, their demand depends on consumer demand for coffee. Currently, consumer demand for coffee has increased because it has become a fashionable beverage, largely due to Starbucks' positioning. Coffee has become an integral part of the lives of professionals, as drinking coffee is a "ritual" for a productive day. Additionally, rising incomes contribute to increased demand for coffee. For example, the growing middle class in China implies an increase in coffee consumption, but this does not always translate to a price increase, as it also depends on supply. Coffee has the peculiarity that price increases do not necessarily affect its consumption level; consumers are willing to buy. Coffee itself is a relatively unstable commodity. Since the demand for coffee is inelastic, slight changes in supply can lead to significant price changes.

Robusta coffee was initially cultivated due to an ecological disaster when coffee production in East Asia was nearly destroyed. Nowadays, Robusta coffee is gaining more relevance because it is less expensive compared to Arabica and has a richer flavor, making it primarily suitable for espresso.

The main producers of Robusta coffee are Vietnam, accounting for 42.3% of the market share, Brazil with 24.6%, and Indonesia contributing 12.85%. Robusta coffee prices fell from 2017 to 2019 due to increased coffee exports from Vietnam. However, prices started to rise from the third quarter of 2019 because Vietnam, the main supplier of Robusta coffee, faced logistical problems during the pandemic, leading to difficulties in delivering goods to ports in China.

In July 2022, Vietnamese coffee exports fell by 17.1% to 113,852 tons. Nevertheless, overall exports grew by 17.3% to 1.13 million tons in the last seven months. According to the U.S. Department of Agriculture reports, they revised their estimates, increasing it from 31.1 tons to 31.58 tons for 2021/2022 but projected a 2.2% decrease for 2022/2023 to 30.9 million tons.

Regarding expectations for Robusta coffee prices by 2025, considering the forecast that the global coffee market will reach $144.68 billion with a growth rate of 7.6% by 2025, it is evident that Robusta coffee production will increase. However, prices are likely to rise in our case due to external factors, such as inflation resulting from a global recession. As coffee is a sensitive commodity, short-term price reductions may occur due to increased production, but in the long run, prices are expected to slowly but consistently increase, possibly by at least 20% from the current price. However, unpredictable events like pandemics or abnormal ultraviolet radiation bursts can impact the market, analogous to unforeseen cases like citrus trees on the Black Sea coast being affected by bugs despite no prior invasions to assess such a risk.